+++ src/models/Discount.php
@@ -11,6 +11,7 @@ use Craft;
 use craft\commerce\base\Model;
 use craft\commerce\Plugin;
 use craft\commerce\records\Discount as DiscountRecord;
+use craft\helpers\ArrayHelper;
 use craft\helpers\UrlHelper;
 use craft\validators\UniqueValidator;
 use DateTime;
@@ -243,7 +244,7 @@ class Discount extends Model
     }
 
     /**
-     * @return array
+     * @return int[]
      */
     public function getCategoryIds(): array
     {
@@ -255,7 +256,7 @@ class Discount extends Model
     }
 
     /**
-     * @return array
+     * @return int[]
      */
     public function getPurchasableIds(): array
     {
@@ -267,7 +268,7 @@ class Discount extends Model
     }
 
     /**
-     * @return array
+     * @return int[]
      */
     public function getUserGroupIds(): array
     {
@@ -281,7 +282,7 @@ class Discount extends Model
     /**
      * Sets the related product type ids
      *
-     * @param array $categoryIds
+     * @param int[] $categoryIds
      */
     public function setCategoryIds(array $categoryIds)
     {
@@ -291,7 +292,7 @@ class Discount extends Model
     /**
      * Sets the related product ids
      *
-     * @param array $purchasableIds
+     * @param int[] $purchasableIds
      */
     public function setPurchasableIds(array $purchasableIds)
     {
@@ -301,7 +302,7 @@ class Discount extends Model
     /**
      * Sets the related user group ids
      *
-     * @param array $userGroupIds
+     * @param int[] $userGroupIds
      */
     public function setUserGroupIds(array $userGroupIds)
     {